+++ consensus/src/state_computer_tests.rs
@@ -207,12 +207,12 @@ async fn commit_should_discover_validator_txns() {
         None,
     );
 
-    // Eventually 4 txns: block metadata, validator txn 0, validator txn 1, state checkpoint.
+    // Eventually 3 txns: block metadata, validator txn 0, validator txn 1.
     let state_compute_result = StateComputeResult::new_dummy_with_compute_status(vec![
             TransactionStatus::Keep(
                 ExecutionStatus::Success
             );
-            4
+            3
         ]);
 
     let blocks = vec![Arc::new(ExecutedBlock::new(
