+++ execution/executor-types/src/lib.rs
@@ -430,6 +430,13 @@ impl StateComputeResult {
         input_txns: Vec<Transaction>,
         block_id: HashValue,
     ) -> Vec<Transaction> {
+        assert_eq!(
+            input_txns.len(),
+            self.compute_status_for_input_txns().len(),
+            "{:?} != {:?}",
+            input_txns.iter().map(|t| t.type_name()).collect::<Vec<_>>(),
+            self.compute_status_for_input_txns()
+        );
         let output = itertools::zip_eq(input_txns, self.compute_status_for_input_txns())
             .filter_map(|(txn, status)| {
                 assert!(
